library & sh

+++ 005.lcd1602_iic/delay.c
@@ -1,13 +1,13 @@
 #include "delay.h"
 
-void delay_us(unsigned int us) {
+void delay_us(uint16_t us) {
     while (us--) {
         _nop_();
     }
 }
 
-void delay_ms(unsigned int ms) {
-    unsigned int i;
+void delay_ms(uint16_t ms) {
+    uint16_t i;
     while (ms--) {
         for (i = 0; i < 101; i++) {
             _nop_();
