+++ lib/delay.h
@@ -2,6 +2,7 @@
 #define __DELAY_H__
 
 #include <intrins.h>
+#include "types.h"
 
 // @11.0592MHz
 #define delay_1us() _nop_()
@@ -9,7 +10,7 @@
 #define delay_3us() { _nop_(); _nop_(); _nop_(); }
 #define delay_5us() { _nop_(); _nop_(); _nop_(); _nop_(); _nop_(); }
 #define delay_10us() { delay_5us(); delay_5us(); }
-void delay_us(unsigned int us);
-void delay_ms(unsigned int ms);
+void delay_us(uint16_t us);
+void delay_ms(uint16_t ms);
 
 #endif
